requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
argon2-cffi==19.2.0
beautifulsoup4==4.6.0     # via wagtail
certifi==2019.11.28       # via requests
cffi==1.13.2              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.0                # via geocoder
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
decorator==4.4.1          # via ratelim
defusedxml==0.6.0         # via python3-openid
django-allauth==0.41.0
django-controlled-vocabulary==0.4.0
django-crispy-forms==1.8.1
django-environ==0.4.5
django-geonames-place==0.4.5
django-kdl-wagtail==0.8.0
django-model-utils==4.0.0
django-modelcluster==5.0.1  # via wagtail
django-polymorphic==2.1.2
django-redis==4.11.0
django-taggit==1.2.0      # via wagtail
django-treebeard==4.3.1   # via wagtail
django==2.2.9
djangorestframework==3.11.0
draftjs-exporter==2.1.7   # via wagtail
edtf==4.0.1
elasticsearch-dsl==7.1.0
elasticsearch==7.5.1      # via elasticsearch-dsl
future==0.18.2            # via geocoder
geocoder==1.38.1          # via django-geonames-place
html5lib==1.0.1           # via wagtail
idna==2.8                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.11.1            # via coreschema
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
pillow==6.2.2             # via wagtail
pycparser==2.19           # via cffi
pyparsing==2.4.6          # via edtf
python-dateutil==2.8.1    # via edtf, elasticsearch-dsl
python-slugify==4.0.0
python3-openid==3.1.0     # via django-allauth
pytz==2019.3
ratelim==0.1.6            # via geocoder
redis==3.3.11
requests-oauthlib==1.3.0  # via django-allauth
requests==2.22.0          # via coreapi, django-allauth, geocoder, requests-oauthlib, wagtail
six==1.14.0               # via argon2-cffi, edtf, elasticsearch-dsl, geocoder, html5lib, python-dateutil, wagtail
sqlparse==0.3.0           # via django
text-unidecode==1.3       # via python-slugify
unidecode==1.1.1          # via wagtail
uritemplate==3.0.1        # via coreapi
urllib3==1.25.8           # via django-controlled-vocabulary, elasticsearch, requests
wagtail==2.7.1
webencodings==0.5.1       # via html5lib
whitenoise==5.0.1
willow==1.3               # via wagtail
